hatchet start lite and hatchet start worker
